Create system test file common.js
