+++ test/functional/system/4.multisig.js
@@ -1,7 +1,5 @@
 'use strict';
 
-var async = require('async');
-var Promise = require('bluebird');
 var chai = require('chai');
 var expect = require('chai').expect;
 var lisk = require('lisk-js');
@@ -10,10 +8,10 @@ var test = require('../../test');
 var _  = test._;
 var accountFixtures = require('../../fixtures/accounts');
 
-var slots = require('../../../helpers/slots');
-
 var application = require('../../common/application');
 var randomUtil = require('../../common/utils/random');
+
+var localCommon = require('./common');
 
 describe('multisignature', function () {
 
@@ -22,7 +20,6 @@ describe('multisignature', function () {
 	before('init sandboxed application', function (done) {
 		application.init({sandbox: {name: 'lisk_test_multisignatures'}}, function (err, scope) {
 			library = scope;
-			console.log('done');
 			done();
 		});
 	});
@@ -31,91 +28,6 @@ describe('multisignature', function () {
 		application.cleanup(done);
 	});
 
-	function forge (cb) {
-		function getNextForger (offset, cb) {
-			offset = !offset ? 1 : offset;
-			var last_block = library.modules.blocks.lastBlock.get();
-			var slot = slots.getSlotNumber(last_block.timestamp);
-			library.modules.delegates.generateDelegateList(function (err, delegateList) {
-				if (err) { return cb (err); }
-				var nextForger = delegateList[(slot + offset) % slots.delegates];
-				return cb(nextForger);
-			});
-		}
-
-		var transactionPool = library.rewiredModules.transactions.__get__('__private.transactionPool');
-		var keypairs = library.rewiredModules.delegates.__get__('__private.keypairs');
-
-		async.waterfall([
-			transactionPool.fillPool,
-			function (cb) {
-				getNextForger(null, function (delegatePublicKey) {
-					cb(null, delegatePublicKey);
-				});
-			},
-			function (delegate, seriesCb) {
-				var last_block = library.modules.blocks.lastBlock.get();
-				var slot = slots.getSlotNumber(last_block.timestamp) + 1;
-				var keypair = keypairs[delegate];
-				test.debug('		Last block height: ' + last_block.height + ' Last block ID: ' + last_block.id + ' Last block timestamp: ' + last_block.timestamp + ' Next slot: ' + slot + ' Next delegate PK: ' + delegate + ' Next block timestamp: ' + slots.getSlotTime(slot));
-				library.modules.blocks.process.generateBlock(keypair, slots.getSlotTime(slot), function (err) {
-					if (err) { return seriesCb(err); }
-					last_block = library.modules.blocks.lastBlock.get();
-					test.debug('		New last block height: ' + last_block.height + ' New last block ID: ' + last_block.id);
-					return seriesCb(err);
-				});
-			}
-		], function (err) {
-			cb(err);
-		});
-	}
-
-	function addTransaction (transaction, cb) {
-		// Add transaction to transactions pool - we use shortcut here to bypass transport module, but logic is the same
-		// See: modules.transport.__private.receiveTransaction
-		library.balancesSequence.add(function (sequenceCb) {
-			library.modules.transactions.processUnconfirmedTransaction(transaction, true, function (err) {
-				if (err) {
-					return setImmediate(sequenceCb, err.toString());
-				} else {
-					return setImmediate(sequenceCb, null, transaction.id);
-				}
-			});
-		}, cb);
-	}
-
-	function addTransactionsAndForge (transactions, cb) {
-		async.waterfall([
-			function addTransactions (waterCb) {
-				async.eachSeries(transactions, function (transaction, eachSeriesCb) {
-					addTransaction(transaction, eachSeriesCb);
-				}, waterCb);
-			},
-			function (waterCb) {
-				setTimeout(function () {
-					forge(waterCb);
-				}, 800);
-			}
-		], function (err) {
-			cb(err);
-		});
-	}
-
-	function getAccountFromDb (address) {
-		return Promise.all([
-			library.db.query('SELECT * FROM mem_accounts where address = \'' + address + '\''),
-			library.db.query('SELECT * FROM mem_accounts2multisignatures where "accountId" = \'' + address + '\''),
-			library.db.query('SELECT * FROM mem_accounts2u_multisignatures where "accountId" = \'' + address + '\'')
-		]).then(function (res) {
-			// Get the first row if resultant array is not empty
-			return {
-				mem_accounts: res[0].length > 0 ? res[0][0] : res[0],
-				mem_accounts2multisignatures: res[1],
-				mem_accounts2u_multisignatures: res[2]
-			};
-		});
-	}
-
 	describe('with LISK sent to multisignature account', function () {
 
 		var multisigAccount;
@@ -123,7 +35,7 @@ describe('multisignature', function () {
 		before('send funds to multisignature account', function (done) {
 			multisigAccount = randomUtil.account();
 			var sendTransaction = lisk.transaction.createTransaction(multisigAccount.address, 1000000000*100, accountFixtures.genesis.password);
-			addTransactionsAndForge([sendTransaction], done);
+			localCommon.addTransactionsAndForge(library, [sendTransaction], done);
 		});
 
 		describe('from multisignature account', function () {
@@ -163,7 +75,7 @@ describe('multisignature', function () {
 					var accountRow;
 
 					before('get mem_account, mem_account2multisignature and mem_account2u_multisignature rows', function () {
-						return getAccountFromDb(multisigAccount.address).then(function (res) {
+						return localCommon.getAccountFromDb(library, multisigAccount.address).then(function (res) {
 							accountRow = res;
 						});
 					});
@@ -264,7 +176,7 @@ describe('multisignature', function () {
 		before('send funds to multisignature account', function (done) {
 			multisigAccount = randomUtil.account();
 			var sendTransaction = lisk.transaction.createTransaction(multisigAccount.address, 1000000000*100, accountFixtures.genesis.password);
-			addTransactionsAndForge([sendTransaction], done);
+			localCommon.addTransactionsAndForge(library, [sendTransaction], done);
 		});
 
 		describe('from multisignature account', function () {
@@ -296,14 +208,14 @@ describe('multisignature', function () {
 
 					multisigTransaction.signatures = [sign1, sign2];
 					multisigTransaction.ready = true;
-					addTransactionsAndForge([multisigTransaction], done);
+					localCommon.addTransactionsAndForge(library, [multisigTransaction], done);
 				});
 
 				describe('sender db rows', function () {
 					var accountRow;
 
 					before('get mem_account, mem_account2multisignature and mem_account2u_multisignature rows', function () {
-						return getAccountFromDb(multisigAccount.address).then(function (res) {
+						return localCommon.getAccountFromDb(library, multisigAccount.address).then(function (res) {
 							accountRow = res;
 						});
 					});
@@ -387,7 +299,7 @@ describe('multisignature', function () {
 						var accountRow;
 
 						before('get mem_account, mem_account2multisignature and mem_account2u_multisignature rows', function () {
-							return getAccountFromDb(multisigAccount.address).then(function (res) {
+							return localCommon.getAccountFromDb(library, multisigAccount.address).then(function (res) {
 								accountRow = res;
 							});
 						});
